Add docker-compose.yml
Running docker commands with args can be a little intimidating for users
who are either new to a CLI and/or Docker.

The Dockerfile has also been updated with a CMD instruction which can be
overiden but allows for simply running 'docker run <image>' or
`docker-compose up`

The README has been updated with instructions on starting and accessing the notebook
container.

Resolves #1